Run previous password attempt in a subshell.
POSIX says that the last stage of a pipeline may or may not be run in
a subshell.  Explicitly wrap the attempt to use the previously entered
password in a subshell to ensure that the script doesn't terminate
with an error if the previously entered password is not correct for
this drive.