Add -c/--config option.
This commit adds a -c / --config option to unlock-cryptroot
so that users may specify additional options for the ssh
connection to their host.  This can be used to add a port
number or any other ssh connection options.  For example:

Host foo.example.net
  Port 2222

This can be used with DROPBEAR_OPTIONS="-p 2222" in
/etc/initramfs-tools/initramfs.conf .  Defaults follow
the same pattern as known_hosts.